refactor: addressed comments on PR
